Refactor preprocessing utils
