+++ config/default.json
@@ -2,7 +2,7 @@
   "dataset": {
     "name": "gqa",
     "split": "train",
-    "version": "all",
+    "version": "balanced",
     "filemap": {
       "root": "./data/gqa"
     }
